tables added

+++ cmds.go
@@ -14,20 +14,30 @@ func exitCmd(params []string) { // args should probably be empty
 	os.Exit(0)
 }
 
-func insertCmd(params []string) { // %d %s, ID Text
+func createTableCmd(params []string) { // %s tableName
+	if len(params) == 1 {
+		if !CreateTable(params[0]) {
+			print("unable to create table, table already exists\n")
+		}
+	} else {
+		print("invalid arg count, 1 expected: %s, table name goes here\n")
+	}
+}
+
+func insertCmd(params []string) { // %s %d %s, tableName ID Text
 	if len(params) >= 2 {
-		rawID, err := strconv.ParseUint(params[0], 10, 32)
+		rawID, err := strconv.ParseUint(params[1], 10, 32)
 		id := uint32(rawID)
 		if err == nil && id != 0 {
-			if !Insert(id, joinOnSpace(params[1:])) {
+			if !Insert(id, joinOnSpace(params[2:]), params[0]) {
 				print("Unable to insert a row with that ID already exists in the DB\n")
 			}
 
 		} else {
-			print("first arg should be <=" + strconv.Itoa(math.MaxUint32) + " for the id of the text\n")
+			print("second arg should be <=" + strconv.Itoa(math.MaxUint32) + " for the id of the text\n")
 		}
 	} else {
-		print("invalid arg count, 2 expected: %d %s, id text goes here\n")
+		print("invalid arg count, 3 expected: %s %d %s ,TableName ID Text\n")
 	}
 }
 
@@ -42,33 +52,35 @@ func joinOnSpace(text []string) string {
 	return textToAddToRow
 }
 
-func selectCmd(params []string) { // %d, ID
-	if len(params) == 1 {
-		rawID, err := strconv.ParseUint(params[0], 10, 32)
+func selectCmd(params []string) { // %s %d, tableName ID
+	if len(params) == 2 {
+		rawID, err := strconv.ParseUint(params[1], 10, 32)
 		id := uint32(rawID)
 		if err == nil && id != 0 {
-			text, found := Select(id)
+			text, found := Select(id, params[0])
 			if found {
 				fmt.Printf("%d: %s\n", id, text)
+			} else {
+				print("text not found\n")
 			}
 		} else {
 			print("could not convert param to int\n")
 		}
 	} else {
-		print("invalid arg count, 1 expected: %d, id \n")
+		print("invalid arg count, 2 expected: %s %d, TableName ID \n")
 	}
 }
 
-func deleteCmd(params []string) { // %d, ID
-	if len(params) == 1 {
-		if params[0] == "database" {
+func deleteCmd(params []string) { // %s %d, tableName ID or %s, tableName
+	if len(params) == 2 {
+		if params[0] == "database" && params[1] == "confirm" {
 			DeleteDB()
 			println("All data in the db and cache has been removed")
 		} else {
-			rawID, err := strconv.ParseUint(params[0], 10, 32)
+			rawID, err := strconv.ParseUint(params[1], 10, 32)
 			id := uint32(rawID)
 			if err == nil && id != 0 {
-				if Delete(id) {
+				if Delete(id, params[0]) {
 					print("row deleted from db\n")
 				} else {
 					print("id not found in memory, row not removed\n")
@@ -78,8 +90,12 @@ func deleteCmd(params []string) { // %d, ID
 				print("could not convert param to int\n")
 			}
 		}
+	} else if len(params) == 1 {
+		if !DeleteTable(params[0]) {
+			print("table not found, not deleted\n")
+		}
 	} else {
-		print("invalid arg count, 1 expected: %d, id \n")
+		print("invalid arg count, 2 expected: %s %d, TableName ID\n")
 	}
 }
 
@@ -88,6 +104,7 @@ func createCmds() map[string]cmdEvent {
 	events["insert"] = insertCmd
 	events["select"] = selectCmd
 	events["delete"] = deleteCmd
+	events["create"] = createTableCmd
 	events[":exit"] = exitCmd
 	return events
 }
